Added files and renamed everything to make more sense as a post feed.
