+++ code/src/components/Header.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Header = ({ title }) => {
+  return (
+    <header>
+      <h1>{title}</h1>
+    </header>
+  )
+};
+
+export default Header;
