+++ code/src/components/NewPost.js
