+++ code/src/components/PostItem.js
@@ -1,0 +1,11 @@
+import React from 'react';
+// import moment from 'moment';
+
+const PostItem = () => {
+  return (
+    <>
+    </>
+  )
+};
+
+export default PostItem;
